requirements.txt
Flask==3.1.0
google-cloud-speech==2.31.0
google-cloud-texttospeech==2.21.0
google-cloud-language==2.11.1
gunicorn==23.0.0